remove circular import, refine design
