+++ apps/common/tests/test_resources.py
@@ -11,7 +11,7 @@ from apps.organizations.models import Organization, RevenueProgram
 from apps.organizations.tests.factories import OrganizationFactory, RevenueProgramFactory
 from apps.pages.models import DonationPage
 from apps.pages.tests.factories import DonationPageFactory
-from apps.users.models import Roles
+from apps.users.choices import Roles
 from apps.users.tests.utils import create_test_user
