+++ apps/organizations/tests/test_org_views.py
@@ -14,7 +14,7 @@ from apps.organizations.tests.factories import (
     RevenueProgramFactory,
 )
 from apps.pages.tests.factories import DonationPageFactory
-from apps.users.models import Roles
+from apps.users.choices import Roles
 from apps.users.tests.utils import create_test_user
 
 
@@ -28,9 +28,7 @@ class OrganizationViewSetTest(AbstractTestCase):
     def setUp(self):
         super().setUp()
         self.list_url = reverse("organization-list")
-        role_assignment_data = {"role_type": Roles.HUB_ADMIN}
-        self.create_user(role_assignment_data=role_assignment_data)
-        self.login()
+        self.set_up_domain_model()
 
     def test_list_of_orgs(self):
         response = self.client.get(self.list_url)
